+++ data/duke.txt
@@ -1,7 +1,5 @@
 D | 0 | return book | 2025-06-06T18:00
 T | 0 | Read book
-D | 0 | Submit report | 2025-08-25T18:00
 E | 0 | Project meeting | 2025-08-26T14:00 | 2025-08-26T16:00
 T | 0 | add this please
 D | 0 | i want this done | 2025-08-24T11:09
-E | 0 | eventA | 2025-08-24T11:43 | 2025-08-25T23:42
