+++ src/main/java/Nami/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package Nami.command;
+import Nami.ui.Ui;
+import Nami.storage.Storage;
+import Nami.task.TaskList;
+import Nami.task.Tasks;
+import Nami.exception.DukeException;
 public class ExitCommand extends Command {
 
     @Override
