+++ src/main/java/Nami/command/MarkCommand.java
@@ -1,3 +1,9 @@
+package Nami.command;
+import Nami.ui.Ui;
+import Nami.storage.Storage;
+import Nami.task.TaskList;
+import Nami.task.Tasks;
+import Nami.exception.DukeException;
 public class MarkCommand extends Command {
     private final int index;
