+++ src/main/java/Nami/exception/DukeException.java
@@ -1,3 +1,4 @@
+package Nami.exception;
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
