+++ src/main/java/Nami/task/Deadlines.java
@@ -1,3 +1,4 @@
+package Nami.task;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
