+++ src/main/java/Nami/ui/Ui.java
@@ -1,3 +1,4 @@
+package Nami.ui;
 import java.util.Scanner;
 
 public class Ui {
